requirement.txt
fastapi
PyJwt
python-dotenv
pydantic
psycopg2-binary
aiosmtplib
asyncio
colorama
bcrypt
uvicorn
schedule
pyodbc
celery